Added MongoDB connection and a basic model
